% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.r
\name{.make.replotdat}
\alias{.make.replotdat}
\title{Convert re.data() output to a dataset for plotting}
\usage{
.make.replotdat(data, var)
}
\arguments{
\item{data}{A data.table object, long format random effects}

\item{var}{A character string, the name of the random effect to plot}
}
\value{
A data.table object with the random effect in a format suitable for plotting
}
\description{
Convert re.data() output to a dataset for plotting
}
\keyword{internal}
