% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranef.r
\name{ranefdata}
\alias{ranefdata}
\title{Create data and plots for brms random effect models}
\usage{
ranefdata(object, usevars, newdata, idvar)
}
\arguments{
\item{object}{a \code{brmsfit} objectx}

\item{usevars}{a character vector of random effects to plot}

\item{newdata}{a data.table object with the data used to generate the random effects, this is used as an anchor for the random intercepts so they have a meaningful 0 point}

\item{idvar}{a character string specifying the grouping variable name for the random effects}
}
\value{
a list with the following components:
\itemize{
  \item{plot}{a list of ggplot objects}
  \item{plotdat}{a list of data.table objects with the data used to generate the plots}
  \item{relong}{a data.table object with the random effects in long format}
  \item{yhat}{a list of data.table objects with the expected values for the random effects}
  \item{usevars}{a character vector of the random effects to plot}
  \item{idvar}{a character string specifying the grouping variable name for the random effects}
}
}
\description{
Create data and plots for brms random effect models
}
